+++ src/ui/circle/bg-dimmer.js
@@ -1,7 +1,7 @@
 function showBackgroundDimmer() {
     backgroundDimmer = document.createElement('div');
     backgroundDimmer.setAttribute('id', 'cmg-page-dimmer');
-    backgroundDimmer.setAttribute('style', `overflow: hidden; z-index: 9999; width:${document.documentElement.clientWidth}px; height: ${document.documentElement.scrollHeight}px;  opacity: 0.0; transition: opacity ${configs.animationDuration}ms ease-out; position:absolute; background: black !important; top: 0px; left: 0px;`);
+    backgroundDimmer.setAttribute('style', `overflow: hidden; z-index: 9999; width:${document.documentElement.clientWidth}px; height: ${document.documentElement.clientHeight}px;  opacity: 0.0; transition: opacity ${configs.animationDuration}ms ease-out; position:fixed; background: black !important; top: 0px; left: 0px;`);
     document.body.appendChild(backgroundDimmer);
 
     setTimeout(function () {